requirements.txt
streamlit>=1.35.0
transformers==4.38.2
torch==2.7.1
Pillow

